+++ jest.config.js
@@ -1,8 +1,8 @@
 module.exports = {
     testEnvironment: 'jsdom',
     // Uncomment the `testMatch` option when we want to run a specific test case
-    testMatch: ['<rootDir>/packages/toolbar/__tests__/testId.test.tsx'],
-    // testRegex: ['(/__tests__/.*|(\\.|/)(test|spec))\\.tsx$'],
+    // testMatch: ['<rootDir>/packages/page-navigation/__tests__/CurrentPageLabelDynamicDocument.test.tsx'],
+    testRegex: ['(/__tests__/.*|(\\.|/)(test|spec))\\.tsx$'],
     moduleNameMapper: {
         '@react-pdf-viewer/core': '<rootDir>/packages/core/src',
         '@react-pdf-viewer/default-layout': '<rootDir>/packages/default-layout/src',