+++ packages/default-layout/__tests__/renderToolbarWithNumberOfPages.test.tsx
@@ -1,0 +1,51 @@
+import * as React from 'react';
+import { render } from '@testing-library/react';
+import { Viewer } from '@react-pdf-viewer/core';
+
+import { mockIsIntersecting } from '../../../test-utils/mockIntersectionObserver';
+import { defaultLayoutPlugin, ToolbarProps, ToolbarSlot } from '../src';
+
+const TestRenderToolbar: React.FC<{
+    fileUrl: Uint8Array;
+}> = ({ fileUrl }) => {
+    const renderToolbar = (Toolbar: (props: ToolbarProps) => React.ReactElement) => (
+        <Toolbar>
+            {(toolbarSlot: ToolbarSlot) => {
+                const { CurrentPageLabel, NumberOfPages } = toolbarSlot;
+                return (
+                    <div data-testid="current-page-label">
+                        <CurrentPageLabel /> of <NumberOfPages />
+                    </div>
+                );
+            }}
+        </Toolbar>
+    );
+
+    const defaultLayoutPluginInstance = defaultLayoutPlugin({
+        renderToolbar,
+    });
+
+    return (
+        <div style={{ height: '50rem', width: '50rem' }}>
+            <Viewer fileUrl={fileUrl} plugins={[defaultLayoutPluginInstance]} />
+        </div>
+    );
+};
+
+test('Test renderToolbar with <NumberOfPages />', async () => {
+    const { findByTestId, getByTestId, queryByTestId } = render(
+        <TestRenderToolbar fileUrl={global['__OPEN_PARAMS_PDF__']} />
+    );
+
+    const viewerEle = getByTestId('core__viewer');
+    mockIsIntersecting(viewerEle, true);
+
+    let pageLabel = await findByTestId('current-page-label');
+    expect(pageLabel.textContent).toEqual('1 of 8');
+
+    // Jump to the fourth page
+    const page = await findByTestId('core__page-layer-3');
+    mockIsIntersecting(page, true);
+    pageLabel = await findByTestId('current-page-label');
+    expect(pageLabel.textContent).toEqual('4 of 8');
+});