+++ CHANGELOG.md
@@ -72,6 +72,7 @@ const thumbnailPluginInstance = thumbnailPlugin({
 -   Navigating between search results should be the same as their appearances on pages
 -   The thumbnails are stuck at loading spinner
 -   The `Cover` component position isn't correct
+-   The number of pages slot `<NumberOfPages />` isn't correct when using the `renderToolbar`
 
 ## v2.11.0